requirements.txt
appdirs==1.4.3
astroid==2.3.2
attrs==19.3.0
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.6.0
pkg-resources==0.0.0
pycrypto==2.6.1
pylint==2.4.3
regex==2019.8.19
requests==2.22.0
rope==0.14.0
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
